+++ src/graphql.rs
@@ -3013,8 +3013,6 @@ pub enum FilterOp {
     ILike,
     RegEx,
     IRegEx,
-    NotRegEx,
-    NotIRegEx
 }
 
 impl ToString for FilterOp {
@@ -3034,8 +3032,6 @@ impl ToString for FilterOp {
             Self::ILike => "ilike",
             Self::RegEx => "regex",
             Self::IRegEx => "iregex",
-            Self::NotRegEx => "nregex",
-            Self::NotIRegEx => "niregex",
         }
         .to_string()
     }
@@ -3060,8 +3056,6 @@ impl FromStr for FilterOp {
             "ilike" => Ok(Self::ILike),
             "regex" => Ok(Self::RegEx),
             "iregex" => Ok(Self::IRegEx),
-            "nregex" => Ok(Self::NotRegEx),
-            "niregex" => Ok(Self::NotIRegEx),
             _ => Err("Invalid filter operation".to_string()),
         }
     }
@@ -3147,8 +3141,6 @@ impl ___Type for FilterTypeType {
                         FilterOp::ILike,
                         FilterOp::RegEx,
                         FilterOp::IRegEx,
-                        FilterOp::NotRegEx,
-                        FilterOp::NotIRegEx
                     ],
                     Scalar::BigInt => vec![
                         FilterOp::Equal,
@@ -3223,9 +3215,7 @@ impl ___Type for FilterTypeType {
                         | FilterOp::Like
                         | FilterOp::ILike
                         | FilterOp::RegEx
-                        | FilterOp::IRegEx
-                        | FilterOp::NotRegEx
-                        | FilterOp::NotIRegEx => __InputValue {
+                        | FilterOp::IRegEx => __InputValue {
                             name_: op.to_string(),
                             type_: __Type::Scalar(scalar.clone()),
                             description: None,