+++ test/sql/string_filters.sql
@@ -109,22 +109,6 @@ begin;
     );
     rollback to savepoint a;
 
-    -- Filter by not regex
-    select jsonb_pretty(
-        graphql.resolve($$
-            {
-              memoCollection(filter: {contents: {nregex: "^F\\w+$"}}) {
-                edges {
-                  node {
-                    contents
-                  }
-                }
-              }
-            }
-        $$)
-    );
-    rollback to savepoint a;
-
     -- iregex is not case sensitive
     select jsonb_pretty(
         graphql.resolve($$
@@ -140,22 +124,5 @@ begin;
         $$)
     );
     rollback to savepoint a;
-
-    -- niregex is not case sensitive
-    select jsonb_pretty(
-        graphql.resolve($$
-            {
-              memoCollection(filter: {contents: {niregex: "^f\\w+$"}}) {
-                edges {
-                  node {
-                    contents
-                  }
-                }
-              }
-            }
-        $$)
-    );
-    rollback to savepoint a;
-
 
 rollback;